refactor: extract AddFoodView, sheets from NutritionView (#431)
NutritionView.swift was 2023 lines. Split into:
- NutritionView.swift (1125 lines) — main view, hero, macro dashboard,
  food log, water card, FAB, data loading
- NutritionSheets.swift (900 lines) — AddFoodView, AlcoholCalculatorView,
  QuickAddView, MacroGoalsSheet

Note: NutritionSheets.swift must be added to the Xcode project.
